requirements.txt
Django==5.1.4
djangorestframework
djangorestframework-simplejwt
django-redis
amqp
asgiref
attrs
billiard
celery
click
click-didyoumean
click-plugins
click-repl
drf-spectacular
Faker
inflection
jsonschema
jsonschema-specifications
kombu
mixer
pillow
prompt_toolkit
psycopg2
PyJWT
python-dateutil
python-decouple
PyYAML
redis
referencing
rpds-py
six
sqlparse
tzdata
uritemplate
validate_email
vine
wcwidth
